function based api design

+++ myproject/myapp/views.py
@@ -1,19 +1,18 @@
 from rest_framework import status
-from django.http import Http404
-from rest_framework.views import APIView 
-from rest_framework.response import Response
+from django.http import JsonResponse
+from rest_framework.decorators import api_view
+from rest_framework.response import Response 
 from . models import Employee
 from . serializers import EmployeeSerializer
 
+@api_view(['GET','POST'])
+def getEmployeeListOrCreate(request):
+	if request.method == "GET":
+		getAllEmployees = Employee.objects.all()
+		serializer = EmployeeSerializer(getAllEmployees, many=True)
+		return Response(serializer.data)
 
-class EmployeeList(APIView):
-
-	def get(self, request):
-			getAllEmployees = Employee.objects.all()
-			serializer = EmployeeSerializer(getAllEmployees, many=True)
-			return Response(serializer.data)
-
-	def post(self, request):
+	elif request.method == "POST":
 		serializer = EmployeeSerializer(data=request.data)
 		if serializer.is_valid():
 			serializer.save()
@@ -21,21 +20,18 @@ class EmployeeList(APIView):
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-class EmployeeDetail(APIView):
+@api_view(['GET','PUT', 'DELETE'])
+def getEmployeeDetailsOrUpdateOrDelete(request, pk):
+	try: 
+		getTheEmployee = Employee.objects.get(id=pk)
+	except Employee.DoesNotExist:
+		return Response(status=status.HTTP_404_NOT_FOUND)
 
-	def get_employee(self, pk):
-		try: 
-			return Employee.objects.get(id=pk)
-		except Employee.DoesNotExist:
-			return Http404
-
-	def get(self, request, pk):
-		getTheEmployee = self.get_employee(pk)
+	if request.method =='GET':
 		serializer = EmployeeSerializer(getTheEmployee, many=False)
 		return Response(serializer.data)
 
-	def put(self, request, pk):
-		getTheEmployee = self.get_employee(pk)
+	elif request.method == 'PUT':
 		serializer = EmployeeSerializer(instance=getTheEmployee,
 			data=request.data)
 		if serializer.is_valid():
@@ -43,8 +39,7 @@ class EmployeeDetail(APIView):
 			return Response(serializer.data)	
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)	
 
-	def delete(self, request, pk):
-		getTheEmployee = self.get_employee(pk)
+	elif request.method == 'DELETE':
 		getTheEmployee.delete()
 		return Response(status=status.HTTP_204_NO_CONTENT)
